more unsigned fixes

+++ src/detail/Chebyshev.h
@@ -220,7 +220,7 @@ template <unsigned int D> struct Chebyshev_Rn {
   // NOTE: valid range of m is 0..n-1
   double_d get_position(const int_d &m) {
     ASSERT((m >= 0).all(), "m should be greater than or equal to 0");
-    ASSERT((m < n).all(), "m should be less than n");
+    ASSERT((m < static_cast<int>(n)).all(), "m should be less than n");
     double_d pos;
     for (size_t d = 0; d < D; ++d) {
       pos[d] = chebyshev_node(m[d], n);
@@ -233,7 +233,7 @@ template <unsigned int D> struct Chebyshev_Rn {
     const unsigned int ii = i * n;
     ASSERT(ii < Sn.size(), "requesting i greater than particles size");
     ASSERT((m >= 0).all(), "m should be greater than or equal to 0");
-    ASSERT((m < n).all(), "m should be less than n");
+    ASSERT((m < static_cast<int>(n)).all(), "m should be less than n");
     double ret = 1.0;
     for (size_t d = 0; d < D; ++d) {
       ret *= Sn[ii + m[d]][d];
@@ -326,7 +326,8 @@ template <unsigned int D> struct ChebyshevRn {
   // NOTE: valid range of m is 0..n-1
   double_d get_position(const int_d &m) {
     ASSERT((m >= 0).all(), "m should be greater than or equal to 0");
-    ASSERT((m < m_Sn.size()).all(), "m should be less than n");
+    ASSERT((m < static_cast<int>(m_Sn.size())).all(),
+           "m should be less than n");
     double_d pos;
     for (size_t d = 0; d < D; ++d) {
       pos[d] = chebyshev_node(m[d], m_Sn.size());
